fix services with exceptions
