+++ practise/pom.xml
@@ -14,7 +14,7 @@
     <name>practise</name>
     <description>practise</description>
     <properties>
-        <java.version>22</java.version>
+        <java.version>21</java.version>
     </properties>
     <dependencies>
         <dependency>
