+++ practise/src/main/java/com/modsen/practise/PractiseApplication.java
@@ -1,0 +1,13 @@
+package com.modsen.practise;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PractiseApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PractiseApplication.class, args);
+    }
+
+}
