+++ practise/src/main/java/com/modsen/practise/service/CategoryService.java
@@ -1,0 +1,14 @@
+package com.modsen.practise.service;
+
+import com.modsen.practise.dto.CategoryDTO;
+
+import java.util.List;
+
+public interface CategoryService {
+    List<CategoryDTO> getAllCategories();
+    CategoryDTO getCategoryById(Long id);
+    CategoryDTO createCategory(CategoryDTO categoryDTO);
+    CategoryDTO updateCategory(Long id, CategoryDTO categoryDTO);
+    void deleteCategory(Long id);
+
+}
