+++ practise/src/main/java/com/modsen/practise/service/ProductService.java
@@ -1,0 +1,14 @@
+package com.modsen.practise.service;
+
+import com.modsen.practise.dto.ProductDTO;
+
+import java.util.List;
+
+public interface ProductService {
+    List<ProductDTO> getAllProducts();
+    ProductDTO getProductById(Long id);
+    ProductDTO createProduct(ProductDTO productDTO);
+    ProductDTO updateProduct(Long id, ProductDTO productDTO);
+    void deleteProduct(Long id);
+    List<ProductDTO> getProductsByCategory(Long categoryId);
+}
