+++ practise/src/main/java/com/modsen/practise/service/UserService.java
@@ -1,0 +1,12 @@
+package com.modsen.practise.service;
+import com.modsen.practise.dto.UserDTO;
+
+import java.util.List;
+
+public interface UserService {
+    List<UserDTO> getAllUsers();
+    UserDTO getUserById(Long id);
+    UserDTO createUser(UserDTO userDTO);
+    UserDTO updateUser(Long id, UserDTO userDTO);
+    void deleteUser(Long id);
+}
