--- practise/src/main/java/org/modsen/practise/PractiseApplication.java
@@ -1,13 +1,0 @@
-package org.modsen.practise;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PractiseApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PractiseApplication.class, args);
-    }
-
-}
