--- practise/src/main/java/org/modsen/practise/controller/OrderController.java
@@ -1,57 +1,0 @@
-package org.modsen.practise.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.modsen.practise.dto.OrderDTO;
-import org.modsen.practise.dto.UserDTO;
-import org.modsen.practise.service.impl.OrderServiceImpl;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/orders")
-public class OrderController {
-
-    private final OrderServiceImpl orderService;
-
-    @GetMapping
-    public ResponseEntity<List<OrderDTO>> getAllOrders() {
-        List<OrderDTO> orders = orderService.getAllOrders();
-        return ResponseEntity.ok(orders);
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<OrderDTO> getOrderById(@PathVariable Long id) {
-        OrderDTO order = orderService.getOrderById(id);
-        return  order != null ? ResponseEntity.ok(order) : ResponseEntity.notFound().build();
-    }
-
-    @PostMapping
-    public ResponseEntity<OrderDTO> createOrder(@RequestBody OrderDTO orderDto) {
-        OrderDTO createdOrder = orderService.createOrder(orderDto);
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdOrder);
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<OrderDTO> updateOrder(@PathVariable Long id, @RequestBody OrderDTO order) {
-        OrderDTO updatedOrder = orderService.updateOrder(id, order);
-        return updatedOrder != null ? ResponseEntity.ok(updatedOrder) : ResponseEntity.notFound().build();
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteOrder(@PathVariable Long id) {
-        orderService.deleteOrder(id);
-        return ResponseEntity.noContent().build();
-    }
-
-//    @GetMapping("/by-user")
-//    public ResponseEntity<List<OrderDTO>> getOrdersByUser(@RequestParam Long userId) {
-//        UserDTO userDTO = new UserDTO();
-//        userDTO.setId(userId);
-//        List<OrderDTO> orders = orderService.getOrdersByUser(userDTO);
-//        return ResponseEntity.ok(orders);
-//    }
-}
