--- practise/src/main/java/org/modsen/practise/dto/CategoryDTO.java
@@ -1,24 +1,0 @@
-package org.modsen.practise.dto;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CategoryDTO {
-    private long id;
-
-    @NotBlank(message = "Name is mandatory")
-    @Size(max = 100, message = "Name should not exceed 100 characters")
-    private String name;
-
-    @Valid
-    private List<ProductDTO> products;
-}
