--- practise/src/main/java/org/modsen/practise/dto/UserDTO.java
@@ -1,32 +1,0 @@
-package org.modsen.practise.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDTO {
-    private long id;
-
-    @Email
-    @Email(message = "Email should be valid")
-    @NotBlank(message = "Email is mandatory")
-    private String email;
-
-    @NotBlank(message = "Login is mandatory")
-    @Size(min = 8, max = 20, message = "Login should be between 3 and 20 characters")
-    private String login;
-
-    @NotBlank(message = "Password is mandatory")
-    @Size(min = 6, message = "Password should be at least 6 characters")
-    private String password;
-
-    private List<OrderDTO> orders;
-}
