--- practise/src/main/java/org/modsen/practise/service/CategoryService.java
@@ -1,14 +1,0 @@
-package org.modsen.practise.service;
-
-import org.modsen.practise.dto.CategoryDTO;
-
-import java.util.List;
-
-public interface CategoryService {
-    List<CategoryDTO> getAllCategories();
-    CategoryDTO getCategoryById(Long id);
-    CategoryDTO createCategory(CategoryDTO categoryDTO);
-    CategoryDTO updateCategory(Long id, CategoryDTO categoryDTO);
-    void deleteCategory(Long id);
-
-}
