--- practise/src/main/java/org/modsen/practise/service/ProductService.java
@@ -1,13 +1,0 @@
-package org.modsen.practise.service;
-
-import org.modsen.practise.dto.ProductDTO;
-
-import java.util.List;
-
-public interface ProductService {
-    List<ProductDTO> getAllProducts();
-    ProductDTO getProductById(Long id);
-    ProductDTO createProduct(ProductDTO productDTO);
-    ProductDTO updateProduct(Long id, ProductDTO productDTO);
-    void deleteProduct(Long id);
-}
