+++ practise/src/test/java/com/modsen/practise/PractiseApplicationTests.java
@@ -1,0 +1,13 @@
+package com.modsen.practise;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PractiseApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
