--- practise/src/test/java/org/modsen/practise/PractiseApplicationTests.java
@@ -1,13 +1,0 @@
-package org.modsen.practise;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PractiseApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
